+++ composer.json
@@ -9,7 +9,7 @@
         "doctrine/doctrine-migrations-bundle": "^3.2",
         "doctrine/orm": "^2.11",
         "sensio/framework-extra-bundle": "^6.2",
-        "stripe/stripe-php": "^7.119",
+        "stripe/stripe-php": "^10.4",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "5.1.*",
         "symfony/console": "5.1.*",
