+++ templates/_partials/_navbar.html.twig
@@ -15,34 +15,24 @@
 			</li>
 
 			<li class="nav-item">
-				<a class="nav-link" href="{{ path('dontpn') }}">Faites un don</a>
+				<a class="nav-link" href="{{ path('don') }}">Faites un don</a>
 			</li>
-			<li class="nav-item  ">
+			<li class="nav-item">
 				<a class="nav-link" href="{{ path('projets') }}">Projets</a>
-<!--
-				<a class="nav-link " href="{{ path('projets') }}" id="navbarDropdown" role="button" data-toggle="dropdown" aria-expanded="false">
-					Projets
-				</a>				
--->
-				<div class="dropdown-menu" aria-labelledby="navbarDropdown">
-					<a class="dropdown-item" href="../debuterSymfony" target="_blank">Postman</a>
-					<a class="dropdown-item" href="" target="_blank">Trocnc</a>
-					<div class="dropdown-divider"></div>
-					<a class="dropdown-item" href="" target="_blank">Module de Paiement Stripe</a>
-				</div>
 			</li>
 			<li class="nav-item">
 				<a class="nav-link" href="{{ path('envoyerMessageContact') }}">Contact</a>
 			</li>
-				{% if app.user %}
-					<li class="nav-item">
-						<a class="nav-link" href="{{ path('security_logout') }}">Logout</a>
-					</li>
-					{% else %}
-					<li class="nav-item">
-						<a class="nav-link" href="{{ path('security_login') }}">Login</a>
-					</li>
-				{% endif %}
+		{% if app.user %}
+			<li class="nav-item">
+				<a class="nav-link" href="{{ path('security_logout') }}">Logout</a>
+			</li>
+			{% else %}
+			<li class="nav-item">
+				<a class="nav-link" href="{{ path('security_login') }}">Login</a>
+			</li>
+			
+		{% endif %}
 
 
 		</ul>
