+++ templates/don.html.twig
@@ -3,18 +3,6 @@
 {% block title %}Faire un don{% endblock %}
 
 {% block body %}
-	<h1>Paie ta dent</h1>
-	<form action="{{ path('charge') }}" method="post" id="payment-form">
-		<div class="form-group">
-			<label for="card-element">
-				Credit or debit card
-			</label>
-			<div
-				id="card-element"><!-- A Stripe Element will be inserted here. -->
-			</div>
-			<!-- Used to display Element errors. -->
-			<div id="card-errors" role="alert"></div>
-		</div>
-		<button class="btn btn-success">Faire une donation</button>
-	</form>
+	<h1>Paiement Stripe</h1>
+	
 {% endblock %}
